requirements-mac.txt
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
asyncpg==0.30.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.1
distro==1.9.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.43
langgraph==0.3.5
langgraph-checkpoint==2.0.18
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.55
langsmith==0.3.13
loguru==0.7.3
mcp==1.5.0
msgpack==1.1.0
nltk==3.9.1
numpy==2.2.3
openai==1.65.5
orjson==3.10.15
packaging==24.2
pgvector==0.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyPDF2==3.0.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
 
setuptools==76.0.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
pypdf==5.4.0
nltk==3.9.1
mcp==1.5.0
python-multipart==0.0.20
pdflatex==0.1.3
docx2pdf==0.1.8
supabase==2.15.0
pillow==11.1.0
google-generativeai==0.8.5
 
pdf2image==1.17.0
pytesseract==0.3.13